pyproject.toml
[project]
name = "steaked-nodes"
description = "A bunch of tools ported from the workshop tools site to ComfyUI nodes along with a few new ones."
version = "1.3.6"
license = {file = "LICENSE"} 
# classifiers = [
#     # For OS-independent nodes (works on all operating systems)
#     "Operating System :: OS Independent",
#
#     # OR for OS-specific nodes, specify the supported systems:
#     "Operating System :: Microsoft :: Windows",  # Windows specific
#     "Operating System :: POSIX :: Linux",  # Linux specific
#     "Operating System :: MacOS",  # macOS specific
#
#     # GPU Accelerator support. Pick the ones that are supported by your extension.
#     "Environment :: GPU :: NVIDIA CUDA",    # NVIDIA CUDA support
#     "Environment :: GPU :: AMD ROCm",       # AMD ROCm support
#     "Environment :: GPU :: Intel Arc",      # Intel Arc support
#     "Environment :: NPU :: Huawei Ascend",  # Huawei Ascend support
#     "Environment :: GPU :: Apple Metal",    # Apple Metal support
# ]

dependencies = ["torch>=2.1.0", "numpy>=1.26.4", "Pillow>=10.0.0"]

[project.urls]
Repository = "https://github.com/StealthNinja1O1/Steaked-nodes"
#  Used by Comfy Registry https://registry.comfy.org
# Documentation = "https://github.com/StealthNinja1O1/Steaked-nodes/wiki"
"Bug Tracker" = "https://github.com/StealthNinja1O1/Steaked-nodes/issues"

[tool.comfy]
PublisherId = "steakedgamer"
DisplayName = "Steaked-nodes"
Icon = ""
includes = [] 
# "requires-comfyui" = ">=1.0.0"  # ComfyUI version compatibility

